Extract location logic to a lib
