+++ app/repos/sequelize/pet.js
@@ -1,28 +1,6 @@
-const Geo = require('geo-nearby');
-
 const config = require('../../../config');
 const models = require('../../models');
-
-function getMargin(pLatitude, pLongitude, km) {
-    const latRadian = (pLatitude * Math.PI) / 180;
-
-    const degLatKm = 110.574235;
-    const degLongKm = 110.572833 * Math.cos(latRadian);
-    const deltaLat = km / degLatKm;
-    const deltaLong = km / degLongKm;
-
-    const topLat = pLatitude + deltaLat;
-    const bottomLat = pLatitude - deltaLat;
-    const leftLng = pLongitude - deltaLong;
-    const rightLng = pLongitude + deltaLong;
-
-    return {
-        topLat,
-        bottomLat,
-        leftLng,
-        rightLng
-    };
-}
+const geo = require('../../libs/geo');
 
 module.exports = {
     findAll(options) {
@@ -62,7 +40,7 @@ module.exports = {
         });
     },
     findNearby(lat, lng, km = config.defaultDistance) {
-        const margin = getMargin(lat, lng, km);
+        const margin = geo.getMargin(lat, lng, km);
 
         return this.findAll({
             where: {
@@ -74,15 +52,7 @@ module.exports = {
                 }
             }
         }).then((pets) => {
-            const geo = new Geo(pets, {
-                setOptions: {
-                    id: 'id',
-                    lat: 'latitude',
-                    lon: 'longitude'
-                }
-            });
-
-            const nearByPetIds = geo.nearBy(lat, lng, km * 1000).map(v => v.i);
+            const nearByPetIds = geo.nearby(pets, lat, lng, km).map(v => v.i);
 
             return pets.filter(pet => nearByPetIds.includes(pet.id));
         });
